Dissolve (haha) the -modules package again
