--- packages/material-composer-modules/README.md
@@ -1,1 +1,0 @@
-# @material-composer/modules
