--- packages/material-composer-modules/package.json
@@ -1,75 +1,0 @@
-{
-  "name": "@material-composer/modules",
-  "private": false,
-  "author": {
-    "name": "Hendrik Mans",
-    "email": "hendrik@mans.de",
-    "url": "https://hendrik.mans.de"
-  },
-  "description": "A collection of Material Composer modules.",
-  "version": "0.0.1",
-  "main": "dist/material-composer-modules.cjs.js",
-  "module": "dist/material-composer-modules.esm.js",
-  "types": "dist/material-composer-modules.cjs.d.ts",
-  "files": [
-    "dist/**",
-    "LICENSE",
-    "README.md"
-  ],
-  "license": "MIT",
-  "sideEffects": false,
-  "scripts": {
-    "clean": "rimraf dist",
-    "dev": "preconstruct watch",
-    "build": "preconstruct build",
-    "release": "yarn ci && yarn publish",
-    "test": "jest --passWithNoTests",
-    "ci": "yarn clean && yarn build && yarn test",
-    "docs": "typedoc src/index.ts"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.18.10",
-    "@babel/preset-env": "^7.18.10",
-    "@babel/preset-react": "^7.17.12",
-    "@babel/preset-typescript": "^7.17.12",
-    "@react-three/fiber": "^8.4.1",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.3.0",
-    "@types/jest": "^28.1.7",
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.5",
-    "@types/three": "^0.143.1",
-    "jest": "^28.1.1",
-    "jest-environment-jsdom": "^28.1.1",
-    "material-composer": "*",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "rimraf": "^3.0.2",
-    "shader-composer": "0.3.0",
-    "three": "^0.143.0",
-    "ts-jest": "^28.0.8",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.3"
-  },
-  "dependencies": {
-    "shader-composer-toybox": "^0.1.1"
-  },
-  "peerDependencies": {
-    "material-composer": "^0.0.1",
-    "shader-composer": "^0.3.0",
-    "three": "^0.143.0"
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env",
-      "@babel/preset-react",
-      [
-        "@babel/preset-typescript",
-        {
-          "isTSX": true,
-          "allExtensions": true
-        }
-      ]
-    ]
-  }
-}
