+++ packages/material-composer/package.json
@@ -12,10 +12,19 @@
   "module": "dist/material-composer.esm.js",
   "types": "dist/material-composer.cjs.d.ts",
   "files": [
+    "units/**",
+    "modules/**",
     "dist/**",
     "LICENSE",
     "README.md"
   ],
+  "preconstruct": {
+    "entrypoints": [
+      "index.ts",
+      "units/index.ts",
+      "modules/index.ts"
+    ]
+  },
   "license": "MIT",
   "sideEffects": false,
   "scripts": {
