+++ packages/material-composer/src/modules/Billboard.ts
@@ -1,4 +1,4 @@
-import { ModuleFactory } from "material-composer"
+import { ModuleFactory } from "."
 import { Billboard as BillboardUnit } from "shader-composer-toybox"
 
 export const Billboard: ModuleFactory = () => (state) => ({
