+++ packages/material-composer/src/modules/Scale.ts
@@ -1,4 +1,4 @@
-import { ModuleFactory } from "material-composer"
+import { ModuleFactory } from "."
 import { Input, Mul } from "shader-composer"
 
 type ScaleProps = {
