+++ packages/material-composer/src/modules/Velocity.ts
